Move ListVirtual to parent
